+++ app/renderers/imgui/help_cache.cpp
@@ -22,7 +22,7 @@ void CachedOperator::build() {
 
         for (const char* param : example_params) {
             Value value = Value::parse(param).value();
-            stack.push_item(StackItem { create_displayables_from(value), std::move(value), false });
+            stack.try_push_item(StackItem { create_displayables_from(value), std::move(value), false });
         }
 
         std::string op_name = filter_name(op.name);