+++ core/error.cpp
@@ -16,6 +16,7 @@ std::ostream& operator<<(std::ostream& os, const Err& error)
 const char* Err::error_strings[] = {
     "Initialization Failure",
     "A specified parameter is invalid",
+    "The stack is full",
     "Load Data Failure",
     "Save Data Failure",
     "Not Implemented",